+++ app/src/main/java/com/example/stocker/repository/baseinterface/BaseCustomerRepository.kt
@@ -6,7 +6,13 @@ import com.example.stocker.pojo.Stock
 interface BaseCustomerRepository {
 
         fun getAllStocks():List<Stock>
-        fun placeOrder(orderId:String,stocks:HashMap<Stock,Int>,stockIds:Array<String>,counts:Array<Int>,total:Long): Pair<Boolean, OrderHistory>
+        fun placeOrder(
+            orderId:String,
+            stocks:HashMap<Stock,Int>,
+            stockIds:Array<String>,
+            counts:Array<Int>,
+            total: Int
+        ): Pair<Boolean, OrderHistory>
         fun getAllOrderHistory():List<OrderHistory>
